[docs] rename main_test.dart to example_test.dart (#96)
It's not safe to name the file `main_test.dart` because the generated test runner file will import that file's main function `as main` which conflicts with the `main` function in the generated test runner file. Took me a few minutes to figure this out when setting up a project.